Split @parallel tests
