+++ test/e2e/specs/specs-calypso/wp-theme-multisite-activate-spec.js
@@ -12,7 +12,6 @@ import * as dataHelper from '../../lib/data-helper';
 
 import ThemeDetailPage from '../../lib/pages/theme-detail-page.js';
 import ThemesPage from '../../lib/pages/themes-page.js';
-import CustomizerPage from '../../lib/pages/customizer-page.js';
 
 import SidebarComponent from '../../lib/components/sidebar-component';
 import SiteSelectorComponent from '../../lib/components/site-selector-component';
@@ -31,69 +30,6 @@ describe( `[${ host }] Themes: All sites (${ screenSize })`, function () {
 	beforeAll( async function () {
 		driver = await driverManager.startBrowser();
 	}, startBrowserTimeoutMS );
-
-	describe( 'Preview a theme @parallel', function () {
-		it( 'Login and select themes', async function () {
-			this.themeSearchName = 'twenty';
-			this.expectedTheme = 'Twenty F';
-
-			this.loginFlow = new LoginFlow( driver, 'multiSiteUser' );
-			await this.loginFlow.loginAndSelectAllSites();
-
-			this.sidebarComponent = await SidebarComponent.Expect( driver );
-			await this.sidebarComponent.selectAllSitesThemes();
-		} );
-
-		it( 'can search for free themes', async function () {
-			this.themesPage = await ThemesPage.Expect( driver );
-			await this.themesPage.waitUntilThemesLoaded();
-			await this.themesPage.showOnlyFreeThemes();
-			await this.themesPage.searchFor( this.themeSearchName );
-
-			await this.themesPage.waitForThemeStartingWith( this.expectedTheme );
-		} );
-
-		describe( 'when a theme more button is clicked', function () {
-			it( 'click theme more button', async function () {
-				await this.themesPage.clickNewThemeMoreButton();
-			} );
-
-			it( 'should show a menu', async function () {
-				const displayed = await this.themesPage.popOverMenuDisplayed();
-				assert( displayed, 'Popover menu not displayed' );
-			} );
-
-			describe.skip( 'when "Try & Customize" is clicked', function () {
-				it( 'click try and customize popover', async function () {
-					await this.themesPage.clickPopoverItem( 'Try & Customize' );
-					this.siteSelector = await SiteSelectorComponent.Expect( driver );
-				} );
-
-				it( 'should show the site selector', async function () {
-					const siteSelectorShown = await this.siteSelector.displayed();
-					return assert( siteSelectorShown, 'The site selector was not shown' );
-				} );
-
-				describe( 'when a site is selected, and Customize is clicked', function () {
-					it( 'select first site', async function () {
-						await this.siteSelector.selectFirstSite();
-						await this.siteSelector.ok();
-					} );
-
-					it( 'should open the customizer with the selected site and theme', async function () {
-						this.customizerPage = await CustomizerPage.Expect( driver );
-						const url = await driver.getCurrentUrl();
-						assert( url.indexOf( this.siteSelector.selectedSiteDomain ) > -1, 'Wrong site domain' );
-						assert( url.indexOf( this.themeSearchName ) > -1, 'Wrong theme' );
-					} );
-
-					afterAll( async function () {
-						await this.customizerPage.close();
-					} );
-				} );
-			} );
-		} );
-	} );
 
 	describe( 'Activate a theme @parallel', function () {
 		it( 'Login and select themes', async function () {
